backend/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.16,<3.17
channels[daphne]>=4.3,<4.4
channels-redis>=4.2,<5.0
django-cors-headers>=4.7,<5.0
psycopg[binary]>=3.2,<4.0
uvicorn[standard]>=0.34,<0.35
